Remove MMXCompatibleModel and make separate function
